# src/app/docs/swagger-suno-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "suno-api",
    "description": "Use API to call the music generation service of Suno.ai and easily integrate it into agents like GPTs.",
    "version": "1.1.0"
  },
...
  "paths": {
    "/api/generate": {
      "post": {
        "summary": "Generate audio based on Prompt.",
        "description": "It will automatically fill in the lyrics.\n\n2 audio files will be generated for each request, consuming a total of 10 credits.\n\n`wait_audio` can be set to API mode:\n\n\u2022 By default, it is set to `false`, which indicates the background mode. It will only return audio task information, and you will need to call the get API to retrieve detailed audio information.\n\n\u2022 If set to `true`, it simulates synchronous mode. The API will wait for a maximum of 100s until the audio is generated, and will directly return the audio link and other information. Recommend using in GPTs and other agents.",
        "tags": ["default"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["prompt", "make_instrumental", "wait_audio"],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "Prompt",
                    "example": "A popular heavy metal song about war, sung by a deep-voiced male singer, slowly and melodiously. The lyrics depict the sorrow of people after the war."
                  },
                  "make_instrumental": {
                    "type": "boolean",
                    "description": "Whether to generate instrumental music",
                    "example": "false"
                  },
                  "model": {
                    "type": "string",
                    "description": "Model name, default is chirp-hawk (v6)",
                    "example": "chirp-hawk|chirp-hawk-wild|chirp-goose"
                  },
                  "wait_audio": {
                    "type": "boolean",
                    "description": "Whether to wait for music generation, default is false, directly return audio task information; set to true, will wait for up to 100s until the audio is generated.",
                    "example": "false"
                  },
                  "weirdness": {
                    "type": "number",
                    "description": "Weirdness slider, 0-100 (default 50, which is omitted from the upstream request).",
                    "example": 80
                  },
                  "style_influence": {
                    "type": "number",
                    "description": "Style Influence slider, 0-100 (default 50, which is omitted from the upstream request).",
                    "example": 30
                  },
                  "variety": {
                    "type": "integer",
                    "description": "Variety slider, integer 0-4 (default 1).",
                    "example": 3
                  },
                  "duration": {
                    "type": "integer",
                    "description": "Fixed song length in seconds, 10-360. Omit for the model default length.",
                    "example": 180
                  },
                  "vocal_gender": {
                    "type": "string",
                    "enum": ["m", "f"],
                    "description": "Constrain the vocal gender. Omit to let the model decide."
                  },
                  "is_max_mode": {
                    "type": "boolean",
                    "description": "Enable Max mode (as on the official Pro client), default false."
                  },
                  "use_personalization": {
                    "type": "boolean",
                    "description": "Personalize the result with the account's 'My Taste' profile, default false."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["0", "1"],
                    "properties": [
                      {
                        "$ref": "#/components/schemas/audio_info"
                      },
                      {
                        "$ref": "#/components/schemas/audio_info"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/custom_generate": {
      "post": {
        "summary": "Generate Audio - Custom Mode",
        "description": "The custom mode enables users to provide additional details about the music, such as music genre, lyrics, and more.\n\n 2 audio files will be generated for each request, consuming a total of 10 credits. \n\n `wait_audio` can be set to API mode:\n\n\u2022 By default, it is set to false, which indicates the background mode. It will only return audio task information, and you will need to call the get API to retrieve detailed audio information.\n\n\u2022 If set to true, it simulates synchronous mode. The API will wait for a maximum of 100s until the audio is generated, and will directly return the audio link and other information. Recommend using in GPTs and other agents.",
        "tags": ["default"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["prompt", "tags", "title"],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "Detailed prompt, including information such as music lyrics.",
                    "example": "[Verse 1]\nCruel flames of war engulf this land\nBattlefields filled with death and dread\nInnocent souls in darkness, they rest\nMy heart trembles in this silent test\n\n[Verse 2]\nPeople weep for loved ones lost\nBattered bodies bear the cost\nSeeking peace and hope once known\nOur grief transforms to hearts of stone\n\n[Chorus]\nSilent battlegrounds, no birds' song\nShadows of war, where we don't belong\nMay flowers of peace bloom in this place\nLet's guard this precious dream with grace\n\n[Bridge]\nThrough the ashes, we will rise\nHand in hand, towards peaceful skies\nNo more sorrow, no more pain\nTogether, we'll break these chains\n\n[Chorus]\nSilent battlegrounds, no birds' song\nShadows of war, where we don't belong\nMay flowers of peace bloom in this place\nLet's guard this precious dream with grace\n\n[Outro]\nIn unity, our strength will grow\nA brighter future, we'll soon know\nFrom the ruins, hope will spring\nA new dawn, we'll together bring"
                  },
                  "tags": {
                    "type": "string",
                    "description": "Music genre",
                    "example": "pop metal male melancholic"
                  },
                  "negative_tags": {
                    "type": "string",
                    "description": "Negative Music genre",
                    "example": "female edm techno"
                  },
                  "title": {
                    "type": "string",
                    "description": "Music title",
                    "example": "Silent Battlefield"
                  },
                  "make_instrumental": {
                    "type": "boolean",
                    "description": "Whether to generate instrumental music",
                    "example": "false"
                  },
                  "model": {
                    "type": "string",
                    "description": "Model name, default is chirp-hawk (v6)",
                    "example": "chirp-hawk|chirp-hawk-wild|chirp-goose"
                  },
                  "wait_audio": {
                    "type": "boolean",
                    "description": "Whether to wait for music generation, default is false, directly return audio task information; set to true, will wait for up to 100s until the audio is generated.",
                    "example": "false"
                  },
                  "weirdness": {
                    "type": "number",
                    "description": "Weirdness slider, 0-100 (default 50, which is omitted from the upstream request).",
                    "example": 80
                  },
                  "style_influence": {
                    "type": "number",
                    "description": "Style Influence slider, 0-100 (default 50, which is omitted from the upstream request).",
                    "example": 30
                  },
                  "variety": {
                    "type": "integer",
                    "description": "Variety slider, integer 0-4 (default 1).",
                    "example": 3
                  },
                  "duration": {
                    "type": "integer",
                    "description": "Fixed song length in seconds, 10-360. Omit for the model default length.",
                    "example": 180
                  },
                  "vocal_gender": {
                    "type": "string",
                    "enum": ["m", "f"],
                    "description": "Constrain the vocal gender. Omit to let the model decide."
                  },
                  "is_max_mode": {
                    "type": "boolean",
                    "description": "Enable Max mode (as on the official Pro client), default false."
                  },
                  "use_personalization": {
                    "type": "boolean",
                    "description": "Personalize the result with the account's 'My Taste' profile, default false."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["0", "1"],
                    "properties": [
                      {
                        "$ref": "#/components/schemas/audio_info"
                      },
                      {
                        "$ref": "#/components/schemas/audio_info"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "audio_info": {
        "type": "object",
        "required": [
          "id",
          "title",
          "image_url",
          "lyric",
          "audio_url",
          "video_url",
          "created_at",
          "model_name",
          "status",
          "gpt_description_prompt",
          "prompt",
          "type",
          "tags"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "audio id"
          },
          "title": {
            "type": "string",
            "description": "music title"
          },
          "image_url": {
            "type": "string",
            "description": "music cover image"
          },
          "lyric": {
            "type": "string",
            "description": "music lyric"
          },
          "audio_url": {
            "type": "string",
            "description": "music download url"
          },
          "video_url": {
            "type": "string",
            "description": "Music video download link, can be used to share"
          },
          "created_at": {
            "type": "string",
            "description": "Create time"
          },
          "model_name": {
            "type": "string",
            "description": "suno model name, chirp-v3"
          },
          "status": {
            "type": "string",
            "description": "The generated states include submitted, queue, streaming, complete."
          },
          "gpt_description_prompt": {
            "type": "string",
            "description": "Simple mode on user input prompt, Suno will generate formal prompts, lyrics, etc."
          },
          "prompt": {
            "type": "string",
            "description": "The final prompt for executing the generation task, customized by the user in custom mode, automatically generated by Suno in simple mode."
          },
          "type": {
            "type": "string",
            "description": "Type"
          },
          "tags": {
            "type": "string",
            "description": "Music genre. User-provided in custom mode, automatically generated by Suno in simple mode."
          }
        },
        "title": "audio_info",
        "description": "Audio Info"
      }
    }
  }
}